Устранила дребезг и внесла правки
