+++ js/advert.js
@@ -1,6 +1,3 @@
-// import {createAdverts} from './data.js';
-
-// const advertInsertBlock = document.querySelector('.map__canvas');
 const advertTemplate = document.querySelector('#card').content.querySelector('.popup');
 
 const TYPES_OF_HOUSES = {
@@ -10,10 +7,6 @@ const TYPES_OF_HOUSES = {
   bungalow: 'Бунгало',
   hotel: 'Отель',
 };
-
-// const similarAdverts = createAdverts();
-// const advertFragment = document.createDocumentFragment();
-
 
 const renderAdvert = function ({author, offer}) {
   const advertElement = advertTemplate.cloneNode(true);
@@ -101,9 +94,6 @@ const renderAdvert = function ({author, offer}) {
   addSrcIfExists(author.avatar, '.popup__avatar');
 
   return advertElement;
-  // advertFragment.appendChild(advertElement);
-  // advertInsertBlock.appendChild(advertFragment);
 };
 
-// const addAdvert = (index) => renderAdvert(similarAdverts[index]);
 export {renderAdvert};
