+++ js/filter.js
@@ -1,4 +1,5 @@
-import {drawAdverts, getOriginalAdverts, setFilteredAdverts} from './map.js';
+import {drawAdverts, getOriginalAdverts, setFilteredAdverts, SIMILAR_ADVERTS_COUNT} from './map.js';
+import {debounce} from './utils/debounce.js';
 
 const mapFilters = document.querySelector('.map__filters');
 const housingTypeFilter = mapFilters.querySelector('#housing-type');
@@ -7,6 +8,10 @@ const roomsFilter = mapFilters.querySelector('#housing-rooms');
 const guestsFilter = mapFilters.querySelector('#housing-guests');
 const featuresFilter = mapFilters.querySelector('#housing-features');
 const features = featuresFilter.querySelectorAll('[name="features"]');
+
+const LOW_PRICE = 10000;
+const HIGH_PRICE = 50000;
+const ROOMS_NOT_FOR_GUESTS = 100;
 
 let selectedHousingType;
 let selectedPrice;
@@ -25,11 +30,11 @@ const filterSimilarAdverts = () => {
     .filter((advert) => {
       if (selectedPrice && selectedPrice !== 'any') {
         if (selectedPrice === 'low') {
-          return advert.offer.price < 10000;
+          return advert.offer.price < LOW_PRICE;
         } else if (selectedPrice === 'middle') {
-          return advert.offer.price >= 10000 && advert.offer.price <= 50000;
+          return advert.offer.price >= LOW_PRICE && advert.offer.price <= HIGH_PRICE;
         } else if (selectedPrice === 'high') {
-          return advert.offer.price > 50000;
+          return advert.offer.price > HIGH_PRICE;
         }
       } else {
         return true;
@@ -45,7 +50,7 @@ const filterSimilarAdverts = () => {
       if (selectedGuests && selectedGuests !== 'any' && selectedGuests !== '0') {
         return advert.offer.guests >= Number(selectedGuests);
       } else if (selectedGuests === '0') {
-        return advert.offer.rooms === 100;
+        return advert.offer.rooms === ROOMS_NOT_FOR_GUESTS;
       } else {
         return true;
       }
@@ -59,9 +64,10 @@ const filterSimilarAdverts = () => {
       }
       return true;
     })
-    .slice(0, 10);
+    .slice(0, SIMILAR_ADVERTS_COUNT);
   setFilteredAdverts(filteredAdverts);
-  drawAdverts();
+  const debouncedDrawAdvert = debounce(drawAdverts);
+  debouncedDrawAdvert();
 };
 
 housingTypeFilter.addEventListener('change', () => {
@@ -82,7 +88,6 @@ roomsFilter.addEventListener('change', () => {
 guestsFilter.addEventListener('change', () => {
   selectedGuests = guestsFilter.value;
   filterSimilarAdverts();
-
 });
 
 featuresFilter.addEventListener('change', () => {
