+++ js/main.js
@@ -1,3 +1,3 @@
 import './map.js';
-import './form-validation.js';
+import './form.js';
 import './filter.js';
